#228 According to code review: Updated comments, fixed coding style, moved test into existing test file.
