+++ src/fparser/common/base_classes.py
@@ -1005,13 +1005,18 @@ class EndStatement(Statement):
         '''Returns a valid Fortran string for this END statement. It
         guarantees that there is no white space after the 'END' in case
         of an unnamed statement.
+
         :param bool isfix: True if the code is in fixed format.
+
+        :returns: the (named or unnamed) valid Fortran END statement \
+                  as a string.
+        :rtype: str
 
         '''
         if self.name:
-            return self.get_indent_tab(isfix=isfix) + 'END %s %s'\
-                   % (self.blocktype.upper(), self.name)
+            return self.get_indent_tab(isfix=isfix) + 'END {0} {1}'\
+                .format(self.blocktype.upper(), self.name)
 
         # Make sure there is no space after an unnamed END:
-        return self.get_indent_tab(isfix=isfix) + 'END %s'\
-            % (self.blocktype.upper())
+        return self.get_indent_tab(isfix=isfix) + 'END {0}'\
+            .format(self.blocktype.upper())
